カメラ、フォントコンポーネントの作成と修正。マルチスレッドでシーン遷移作成

+++ my_framework/Src/Framework/Input/Keyboard.cpp
@@ -2,36 +2,36 @@
 #include "../../../environment.h"
 
 
-BYTE Keyboard::GetTrgKeyType() {
+int Keyboard::GetTrgKeyType() {
 	for (int i = 0x01; i < 0xdf; i++) {
 		if (Trg(i)) return i;
 	}
 	return 0;
 }
-BYTE Keyboard::GetOnKeyType() {
+int Keyboard::GetOnKeyType() {
 	for (int i = 0x01; i < 0xdf; i++) {
 		if (On(i)) return i;
 	}
 	return 0;
 }
-BYTE Keyboard::GetRelKeyType() {
+int Keyboard::GetRelKeyType() {
 	for (int i = 0x01; i < 0xdf; i++) {
 		if (Rel(i)) return i;
 	}
 	return 0;
 }
-bool Keyboard::On(BYTE key) {
+bool Keyboard::On(int key) {
 	if (key > 223 || key < 0) return false;
 	BYTE* cur = DirectInput::GetCurrentKeyboard();
 	return (cur[key] != 0);
 }
-bool Keyboard::Trg(BYTE key) {
+bool Keyboard::Trg(int key) {
 	if (key > 223 || key < 0) return false;
 	BYTE* cur = DirectInput::GetCurrentKeyboard();
 	BYTE* old = DirectInput::GetOldKeyboard();
 	return (cur[key] != 0 && old[key] == 0);
 }
-bool Keyboard::Rel(BYTE key) {
+bool Keyboard::Rel(int key) {
 	if (key > 223 || key < 0) return false;
 	BYTE* cur = DirectInput::GetCurrentKeyboard();
 	BYTE* old = DirectInput::GetOldKeyboard();
