+++ my_framework/Src/Framework/Input/Keyboard.h
@@ -9,10 +9,10 @@
 class Keyboard {
 public:
 	//キーボード(DIK)
-	static BYTE GetTrgKeyType(); //押されたキーのDirectInputキー識別コードの種類を変えす
-	static BYTE GetOnKeyType(); //押されたキーのDirectInputキー識別コードの種類を変えす
-	static BYTE GetRelKeyType(); //押されたキーのDirectInputキー識別コードの種類を変えす
-	static bool On(BYTE key);
-	static bool Trg(BYTE key);
-	static bool Rel(BYTE key);
+	static int GetTrgKeyType(); //押されたキーのDirectInputキー識別コードの種類を変えす
+	static int GetOnKeyType(); //押されたキーのDirectInputキー識別コードの種類を変えす
+	static int GetRelKeyType(); //押されたキーのDirectInputキー識別コードの種類を変えす
+	static bool On(int key);
+	static bool Trg(int key);
+	static bool Rel(int key);
 };
